requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-allauth==65.3.1
django-appconf==1.0.6
django-filebrowser==4.0.3
django-filter==24.3
django-grappelli==3.0.9
django-imagekit==5.0.0
django-js-asset==3.0.1
django-tinymce==4.1.0
kombu==5.4.2
pilkit==3.0
pillow==11.1.0
pip==25.0
prompt_toolkit==3.0.50
psycopg==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13